// src/api-services/models/create-item-sale-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 规范化接口演示
 * 让 .NET 开发更简单，更通用，更流行。
 *
 * OpenAPI spec version: 1.0.0
 * Contact: monksoul@outlook.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateItemSaleDto
 */
export interface CreateItemSaleDto {
    /**
     * 
     * @type {string}
     * @memberof CreateItemSaleDto
     */
    batchNumber?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof CreateItemSaleDto
     */
    batchDate?: Date | null;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    quantity: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    receivedQuantity: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    unitPriceNet: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    discount: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    taxRate: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    tax: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    total: number;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    itemId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    itemBatchId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    saleId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CreateItemSaleDto
     */
    unitPriceId: number;
}